check_rabbitmq-ack-rate: change logic + fix a couple of bugs
The biggest issue is that the check compares an ack rage and a get rate. There
should not be difference between those values and it can only alerts when
a processing client would be silly. Then the messages would be fetched from RabbitMQ
but not acked and the ratio would grow. But the get and ack rate are divided and
the result is compared with default options to 0.3 (crit) or 0.5 (warn). The ratio between
those rates can never be lower than 1 (a client cannot acks faster than it's getting
messages from RabbitMQ) so it cannot work anyway.

This check is replaced by a simple comparison of the ack rate with a static value
so the check alerts when a processing is slow.

The minor bugs and improvements:

- fix typo: '$echo' > '$( echo'
- fix typo 'queue_processing_rate_cIritical'
- do not run a superfluous subshell when storing a message into a 'message' variable
- remove a superflous quotes